+++ apps/admin/src/features/heroImagesManager/components/HeroImageCreator.tsx
@@ -2,7 +2,7 @@ import { Button, Modal } from 'antd'
 
 import { HeroImageForm } from './HeroImageForm'
 
-import { useModal } from '~/hooks/useModal'
+import { useModal } from '~/shared/hooks/useModal'
 import { MESSAGES } from '../constant/constants'
 
 export const HeroImageCreator = () => {
