+++ apps/admin/src/routes/notice/$postId.tsx
@@ -1,7 +1,7 @@
 import { createFileRoute, useMatch } from '@tanstack/react-router'
 import { usePostServiceGetApiV1PostsByPostIdSuspense } from '~/apis/community/queries/suspense'
-import { PATH } from '~/constants/path'
 import { Board } from '~/shared/components/Post'
+import { PATH } from '~/shared/constants/path'
 
 export const Route = createFileRoute('/notice/$postId')({
   component: PostDetailPage,
