--- apps/admin/src/shared/hooks/Post/index.ts
@@ -1,4 +1,0 @@
-export { useDeletePost } from './useDeletePost'
-export { usePostDetail } from './usePostDetail'
-export { useCreatePost } from './useCreatePost'
-export { usePatchPost } from './usePatchPost'
