+++ apps/admin/src/shared/hooks/useRefreshToken.ts
@@ -1,5 +1,5 @@
 import { useMutation } from '@tanstack/react-query'
-import { END_POINT } from '~/constants/api'
+import { END_POINT } from '~/shared/constants/api'
 import { getToken } from '~/shared/utils/api'
 import { authServices } from '~/shared/utils/auth'
 import { authHttp } from '~/shared/utils/http'
