+++ apps/admin/src/shared/hooks/useSignIn.ts
@@ -1,6 +1,6 @@
 import { useMutation } from '@tanstack/react-query'
 import { useNavigate } from '@tanstack/react-router'
-import { END_POINT } from '~/constants/api'
+import { END_POINT } from '~/shared/constants/api'
 import { authServices } from '~/shared/utils/auth'
 import { authHttp } from '~/shared/utils/http'
