+++ apps/admin/src/shared/utils/auth.ts
@@ -1,9 +1,9 @@
 import { useNavigate } from '@tanstack/react-router'
 
-import { ACCESS_TOKEN_KEY, REFRESH_TOKEN_KEY } from '~/constants/api'
+import { ACCESS_TOKEN_KEY, REFRESH_TOKEN_KEY } from '~/shared/constants/api'
 import { decodeJwt, removeTokens } from '~/shared/utils'
 
-import type { Tokens } from '~/hooks/useSignIn'
+import type { Tokens } from '~/shared/hooks/useSignIn'
 
 const authServices = () => {
   const navigate = useNavigate()
